+++ interface_stone_smart_flutter/lib/interface_stone_smart_flutter.dart
@@ -1,9 +1,24 @@
 library interface_stone_smart_flutter;
 
-export 'enum/printer_type.dart';
+// Types
+export 'utils/types/payment_type_call.dart';
+export 'utils/types/payment_type_credit.dart';
+export 'utils/types/payment_type_handler.dart';
+export 'utils/types/payment_type_transaction.dart';
+export 'utils/types/printer_type.dart';
+
+// Extensions
+export 'utils/extensions/payment_type_call_ext.dart';
+export 'utils/extensions/payment_type_credit_ext.dart';
+export 'utils/extensions/payment_type_handler_ext.dart';
+export 'utils/extensions/payment_type_transaction_ext.dart';
+export 'utils/extensions/printer_type_ext.dart';
+
+// Handlers
 export 'handler/istone_handler.dart';
 export 'handler/istone_smart_handler.dart';
+
+// Models
 export 'model/stone_printer.dart';
 export 'model/stone_response.dart';
 export 'model/stone_transaction_model.dart';
-export 'utils/payment_types.dart';