+++ interface_stone_smart_flutter/lib/utils/types/payment_type_credit.dart
@@ -1,0 +1,4 @@
+enum PaymentTypecredit {
+  salesman,
+  client,
+}