+++ interface_stone_smart_flutter/lib/utils/types/payment_type_handler.dart
@@ -1,0 +1,14 @@
+enum PaymentTypeHandler {
+  onTransactionSuccess,
+  onError,
+  onMessage,
+  onLoading,
+  writeToFile,
+  onAbortedSuccessfully,
+  disposeDialog,
+  activeDialog,
+  onAuthProgress,
+  onTransactionInfo,
+  onFinishedResponse,
+  onChanged
+}