+++ interface_stone_smart_flutter/lib/utils/types/printer_type.dart
@@ -1,0 +1,4 @@
+enum PrinterType {
+  text,
+  image;
+}