+++ stone_smart_flutter/android/src/main/java/com/qztech/stone_smart_flutter/payments/PaymentsUseCase.java
@@ -66,19 +66,19 @@ public class PaymentsUseCase {
                               int parc,
                               boolean withInterest,
                               Map<StoneKeyType, String> stoneKeys,
-                              boolean isPrinter
+                              boolean printCustomerSlip
                               ){
     isAbortRunning = false;
     if(stoneKeys == null) {
       checkUserModel(context);
-      transaction(context, amount, typeTransaction, initiatorTransactionKey, parc, withInterest, isPrinter);
+      transaction(context, amount, typeTransaction, initiatorTransactionKey, parc, withInterest, printCustomerSlip);
       return;
     }
     if(userModel == null) {
       userModel = StoneStart.init(context, stoneKeys);
     }
     
-    transaction(context, amount, typeTransaction, initiatorTransactionKey, parc, withInterest, isPrinter);
+    transaction(context, amount, typeTransaction, initiatorTransactionKey, parc, withInterest, printCustomerSlip);
   }
 
   private void checkUserModel(Context context) {
@@ -123,7 +123,7 @@ public class PaymentsUseCase {
     mFragment.onMessage("Transação concluída");
   }
 
-  public void handleTransactionSuccess(Context context, ActionResult actionResult, boolean isPrinter, TransactionObject transaction) {
+  public void handleTransactionSuccess(Context context, ActionResult actionResult, boolean printCustomerSlip, TransactionObject transaction) {
     TransactionDAO transactionDAO = new TransactionDAO(context);
     TransactionObject transactionObject = transactionDAO.findTransactionWithInitiatorTransactionKey(transaction.getInitiatorTransactionKey());
     if(!userModel.isEmpty()){
@@ -131,7 +131,6 @@ public class PaymentsUseCase {
       actionResult.setUserModel(userModelString);
     }
 
-    boolean isPaymentApproved = false;
     TransactionStatusEnum transactionStatus = TransactionStatusEnum.UNKNOWN;
     TransactionStatusEnum currentTransactionStatus = TransactionStatusEnum.UNKNOWN;
 
@@ -144,10 +143,8 @@ public class PaymentsUseCase {
       actionResult.setTransactionStatus(transactionStatus.toString());
       actionResult.setMessageFromAuthorize(posTransactionProvider.getMessageFromAuthorize());
       actionResult.setAuthorizationCode(posTransactionProvider.getAuthorizationCode());
-      if (isPrinter && currentTransactionObject != null) {
-        printerReceiptTransaction(context, currentTransactionObject, transactionStatus);
-      }
-      isPaymentApproved = transactionStatus == TransactionStatusEnum.APPROVED || currentTransactionStatus == TransactionStatusEnum.APPROVED;
+      printerReceiptTransaction(context, currentTransactionObject, transactionStatus, printCustomerSlip);
+      boolean isPaymentApproved = transactionStatus == TransactionStatusEnum.APPROVED || currentTransactionStatus == TransactionStatusEnum.APPROVED;
       actionResult.buildResponseStoneTransaction(transactionObject, isPaymentApproved);
       checkStatusWithErrorTransaction(transactionStatus, context);
     }
@@ -184,10 +181,10 @@ public class PaymentsUseCase {
           Context context,
           String amount,
           int typeTransaction,
-          String initiatorTransactionKey,
+          String InitiatorTransactionKey,
           int parc,
           boolean withInterest,
-          boolean isPrinter
+          boolean printCustomerSlip
   ) {
 
     BasicResult basicResult = new BasicResult();
@@ -207,10 +204,10 @@ public class PaymentsUseCase {
       currentTransactionObject = transaction;
 
       // Essa validação é para verificar se a transação já foi realizada (recomendação da própria Stone) para evitar duplicidade
-      if(initiatorTransactionKey != null && !initiatorTransactionKey.isEmpty()){
+      if(InitiatorTransactionKey != null && !InitiatorTransactionKey.isEmpty()){
         TransactionDAO transactionDAO = new TransactionDAO(context);
 
-        TransactionObject transactionObject = transactionDAO.findTransactionWithInitiatorTransactionKey(initiatorTransactionKey);
+        TransactionObject transactionObject = transactionDAO.findTransactionWithInitiatorTransactionKey(InitiatorTransactionKey);
         if(transactionObject != null) {
           if(transactionObject.getTransactionStatus() == TransactionStatusEnum.APPROVED) {
             actionResult.setTransactionStatus(transactionObject.getTransactionStatus().toString());
@@ -236,7 +233,7 @@ public class PaymentsUseCase {
           return;
         }
 
-        transaction.setInitiatorTransactionKey(initiatorTransactionKey);
+        transaction.setInitiatorTransactionKey(InitiatorTransactionKey);
       }
 
 
@@ -248,7 +245,7 @@ public class PaymentsUseCase {
       posTransactionProvider.setConnectionCallback(new StoneActionCallback() {
         @Override
         public void onSuccess() {
-          handleTransactionSuccess(context, actionResult, isPrinter, transaction);
+          handleTransactionSuccess(context, actionResult, printCustomerSlip, transaction);
         }
         @Override
         public void onStatusChanged(Action action) {
@@ -286,25 +283,47 @@ public class PaymentsUseCase {
     mFragment.onFinishedResponse(convertBasicResultToJson(basicResult));
   }
 
-  public void printerCurrentTransaction(Context context, boolean isPrinter) {
-    if(currentTransactionObject == null || !isPrinter) return;
+  public void printerCurrentTransaction(Context context, boolean printCustomerSlip) {
+    if(currentTransactionObject == null) {
+      Log.d("print", "****ERROR_currentTransactionObject: invalid null value");
+      return;
+    }
+
     try {
       mStonePrinter.printerFromTransaction(context, currentTransactionObject);
     } catch (Exception error) {
-      Log.d("print", "****ERRO_printerCurrentTransaction: " + error.getMessage());
+      Log.d("print", "****ERROR_printerCurrentTransaction: " + error.getMessage());
     }
   }
 
-  public void printerReceiptTransaction(Context context, TransactionObject transactionObject,TransactionStatusEnum status ) {
+  public void printerReceiptTransaction(
+          Context context,
+          TransactionObject transactionObject,
+          TransactionStatusEnum status,
+          boolean printCustomerSlip
+  ) {
     try {
-      boolean isPrinterValid = status == TransactionStatusEnum.APPROVED || transactionObject.getTransactionStatus() == TransactionStatusEnum.APPROVED;
+      boolean isValidTransaction = status == TransactionStatusEnum.APPROVED || transactionObject.getTransactionStatus() == TransactionStatusEnum.APPROVED;
+      
+      if(!isValidTransaction) return;
+      printReceipt(context, transactionObject, ReceiptType.MERCHANT);
+      Log.d("print", "****printMerchantVia");
 
-        if(isPrinterValid) {
-        PosPrintReceiptProvider printer = new PosPrintReceiptProvider(context, transactionObject, ReceiptType.MERCHANT);
-        printer.execute();
-      }
+      if (!printCustomerSlip) return;
+      printReceipt(context, transactionObject, ReceiptType.CLIENT);
+      Log.d("print", "****printCustomerSlip ");
+
     } catch(Exception e) {
-      Log.d("print", "****ERRO_printerReceiptTransaction: " + e.getMessage());
+      Log.d("print", "****ERROR_printerReceiptTransaction: " + e.getMessage());
+    }
+  }
+
+  public void printReceipt(Context context, TransactionObject transactionObject, ReceiptType type) {
+    try {
+      PosPrintReceiptProvider printer = new PosPrintReceiptProvider(context, transactionObject, type);
+      printer.execute();
+    } catch (Exception e) {
+      Log.d("print", "****ERROR_printMerchantVia: " + e.getMessage());
     }
   }
 
@@ -316,7 +335,7 @@ public class PaymentsUseCase {
       posTransactionProvider.abortPayment();
       isAbortRunning = true;
     } catch (Exception error){
-      Log.d("print", "****ERRO_abortCurrentPosTransaction: " + error.getMessage());
+      Log.d("print", "****ERROR_abortCurrentPosTransaction: " + error.getMessage());
       BasicResult basicResult = new BasicResult();
       basicResult.setResult(999999);
       basicResult.setErrorMessage(error.getMessage());
@@ -547,7 +566,7 @@ public class PaymentsUseCase {
   public void checkStatusWithErrorTransaction(TransactionStatusEnum status, Context context) {
     if(status == null || (status != TransactionStatusEnum.WITH_ERROR && status != TransactionStatusEnum.APPROVED) ) {
       if(posTransactionProvider != null &&  !isAbortRunning) {
-        mFragment.onMessage("Abortando a transacao");
+        mFragment.onMessage(posTransactionProvider.getMessageFromAuthorize());
         posTransactionProvider.abortPayment();
         isAbortRunning = true;
       }
@@ -555,7 +574,7 @@ public class PaymentsUseCase {
     }
 
      if(status == TransactionStatusEnum.WITH_ERROR) {
-       mFragment.onMessage("Revertendo a transacao");
+       mFragment.onMessage("Revertendo a transação");
        onReversalTransaction(context);
      }
   }
@@ -572,6 +591,7 @@ public class PaymentsUseCase {
           basicResult.setResult(0);
           basicResult.setMessage("OK");
           mFragment.onFinishedResponse(convertBasicResultToJson(basicResult));
+          printReceipt(context, currentTransactionObject, ReceiptType.MERCHANT);
           // Log.d("print", "****************************************************************** ABORT PIX 7 ******************************************************************");
         }
 
@@ -594,7 +614,7 @@ public class PaymentsUseCase {
     }
   }
 
-    public void getTransactionByInitiatorTransactionKey(Context context, String initiatorTransactionKey) {
+    public void getTransactionByInitiatorTransactionKey(Context context, String InitiatorTransactionKey) {
       BasicResult basicResult = new BasicResult();
       basicResult.setMethod("paymentGetTransactionByInitiatorTransactionKey");
       ActionResult actionResult = new ActionResult();
@@ -602,7 +622,7 @@ public class PaymentsUseCase {
 
       try {
         TransactionDAO transactionDAO = new TransactionDAO(context);
-        TransactionObject transactionObject = transactionDAO.findTransactionWithInitiatorTransactionKey(initiatorTransactionKey);
+        TransactionObject transactionObject = transactionDAO.findTransactionWithInitiatorTransactionKey(InitiatorTransactionKey);
 
         if (transactionObject == null) {
           handleTransactionError(context, "Transação não encontrada", actionResult, basicResult);
@@ -624,11 +644,11 @@ public class PaymentsUseCase {
           String userModelString = getGson().toJson(userModel.get(0));
           actionResult.setUserModel(userModelString);
         }
-
+        actionResult.buildResponseStoneTransaction(transactionObject, isPaymentApproved);
         actionResult.buildResponseStoneTransaction(transactionObject, isPaymentApproved);
 
-        actionResult.buildResponseStoneTransaction(transactionObject, isPaymentApproved);
         String jsonStoneResult = convertActionToJson(actionResult);
+
         finishTransaction(jsonStoneResult);
 
       } catch (Exception error) {
@@ -641,9 +661,7 @@ public class PaymentsUseCase {
       ActionResult actionResult = new ActionResult();
       try {
         mFragment.onMessage("Obtendo todas as transações");
-        // acessa todas as transacoes do banco de dados
         TransactionDAO transactionDAO = new TransactionDAO(context);
-        // cria uma lista com todas as transacoes
         transactionObjects = transactionDAO.getAllTransactionsOrderByIdDesc();
 
 
@@ -674,7 +692,11 @@ public class PaymentsUseCase {
       mStonePrinter.customPrinter(params, context);
     }
 
-    public void printerFromBase64(String params, Context context) {
-      mStonePrinter.printerFromBase64(params, context);
+    public void printFromBase64(String params, Context context) {
+      mStonePrinter.printFromBase64(params, context);
+    }
+
+    public void printWrapPaper(int lines, Context context) {
+      mStonePrinter.printWrapPaper(lines, context);
     }
 }